Variables storing credentials are overwritten after use
