+++ src/metro/syncing.cpp
@@ -107,81 +107,6 @@ namespace metro {
         return diff.num_deltas() == 0;
     }
 
-    void get_keys(string *pub, string *pri) {
-#ifdef _WIN32
-        cout << "Metro currently doesn't support SSH on Windows. Please use HTTPS." << endl;
-        return;
-#elif __unix__
-        const char* home;
-        home = getenv("HOME");
-
-        if (home == NULL) {
-            home = getpwuid(getuid())->pw_dir;
-        }
-
-        ifstream keyfile;
-        keyfile.open(string(home) + "/.ssh/id_rsa.pub");
-        if (keyfile.is_open()) {
-            string line;
-            while (getline(keyfile, line)) {
-                pub->append(line);
-            }
-            keyfile.close();
-        } else {
-            cout << "Public Key not found at " + string(home) + "/.ssh/id_rsa.pub" << endl;
-        }
-        keyfile.open(string(home) + "/.ssh/id_rsa");
-        if (keyfile.is_open()) {
-            string line;
-            while (getline(keyfile, line)) {
-                pri->append(line);
-            }
-            keyfile.close();
-        } else {
-            cout << "Public Key not found at " + string(home) + "/.ssh/id_rsa" << endl;
-        }
-#endif //_WIN32
-    }
-
-    int acquire_credentials(git_cred **cred, const char *url, const char *username_from_url, unsigned int allowed_types, void *payload) {
-        int err = GIT_OK;
-        auto credentials = static_cast<git_cred**>(payload);
-
-        if (*credentials == nullptr) {
-            string username;
-            string password;
-            switch (allowed_types) {
-                case GIT_CREDTYPE_DEFAULT:
-                    err = git_cred_default_new(credentials);
-                    break;
-                case GIT_CREDTYPE_USERPASS_PLAINTEXT:
-                    cout << "Username for " << url << ": ";
-                    getline(cin, username);
-                    cout << "Password for " << username << ": ";
-                    password = read_password();
-
-                    err = git_cred_userpass_plaintext_new(credentials, username.c_str(), password.c_str());
-                    break;
-                default:
-                    cout << "Username for " << url << ": ";
-                    getline(cin, username);
-                    cout << "SSH Keystore Password: ";
-                    password = read_password();
-
-                    string pub, pri;
-                    get_keys(&pub, &pri);
-//                cout << "Metro currently doesn't support SSH. Please use HTTPS." << endl;
-                    cout << "Public key is:\n" << pub << endl;
-                    cout << "Private key is:\n" << pri << endl;
-                    err = git_cred_ssh_key_new(credentials, username.c_str(), pub.c_str(), pri.c_str(), password.c_str());
-                    break;
-            }
-        }
-
-        *cred = *credentials;
-        return err;
-    }
-
     void clear_sync_cache(const Repository& repo) {
         repo.foreach_reference([](const Branch& ref, const void *payload) {
             if (has_prefix(ref.reference_name(), "refs/synced/")) {
